requirements.txt
flask==2.3.3
Werkzeug==2.3.7
google-generativeai
python-dotenv
supabase
